Explicit Error handling
This change introduces some explicit error handling to the retrieval
and execution of nmpi jobs. More specifically we explicitely try to
handle all possible errors of the preparatory steps before saga
submission and kill the nmpi job if this happens.

Furthermore we are also waiting for the completion of jobs differently.
Even though we still block until all jobs retrieved from the server have
completed or thrown an error, we now at least send completion notifications
to the nmpi server out of order.

Finally we've also associated the start_time field to individual saga_jobs
and not the JobRunner object, which now enables correct processing of the still
fragile output file handling based on time of creation.